Websocket: wrap database transactions in try catch so we can debug dbus service errors
Avoids unpleasant situations were the websocket server might be under
load and bail out - actually stopping the service entirely.

Wrap also dead job detection in a database transaction.

Even tho we observed that auto_duplicate in production caused the timer
to kill completely the dbus service, in this commit we won't disable it
until we can confirm it's the root of the problem
.
Important: we need to depend on new Net::DBus 1.1