Better reflect job execution through states
With all the STATES we have (and those forthcomming), we need to
know better what states belong to whic slice of it's lifetime.

Remove hardcoded states

Move Assigned back to EXECUTION

Right now, a job that is assigned is a job that is "pending" execution
For the time being we still have some jobs that spend some time here,
this will allow to easily restart as 4eb1114f did.