Refactor worker status messages
That code already called for refactoring since it was duplicated in the codebase, so let's make the code more readable as well and add unit test for checking message
construction.

This commits also remove the stop_job event subscription to notify
immediately that the worker is free, leaving that to the timer.
That would give also a bit more time before picking up a new job.