Multiple loggings channels
Sometimes it's required that we are able to log text to multiple
files. Currently the default OpenQA::Setup::app only logs to the
destination specified in the config files.

With these changes if the developer wants or needs to log to
another file/destination, just needs to call the normal logging
functions with an extra parameter, which is the name of the
"channel". Example:

> log_debug($message, 'this is name of channel');

This will work as long as the "channel" already exists. Otherwise,
it will use the default logging mechanism.

To create a new "channel" just invoke add_log_channel with the
**id** as first parameter. The rest of the parameters are the
same as the ones required for Mojo::Log. Example:

> add_log_channel('channel name', path=>'test.log', level=>'debug')
> unless get_channel_handle('autoinst');

This will create a Mojo::Log object that will be associated with
the name 'channel name'