Use travis stages to reduce failing tests in CI
While we benefit a lot by tests that are running in parallel, very
often, some of this tests (Mainly FULL_STACK and SCHEDULER_FULLSTACK)
tend to fail, a higher sucess rate can be seen when these are restarted
separately.

While it's a Travi's beta feature, looks robust enough and improved
since last time I tried this.