Use OPENQA_TEST_IPC for 12-admin-workers.t
A test recently added by @coolo causes it to hit DBus, so we
need this or the test fails in package build environments. The
call that gets made is `ws_is_worker_connected`, so also set
up the test WebSockets service. The test will actually still
pass even if the calls fail, but let's be 'correct'.