`limit_assets`: remove asset that breaks limit, not next one (#1536)
At present, when looping through the assets for a given job
group, we decide whether to make the asset a deletion candidate
*before* we count its size against the limits. So effectively
we keep the minimum amount of assets needed to *exceed* the
size limit, rather than the maximum amount of assets that can
fit within it. This is clearly incorrect. We can fix it by
counting the asset's size against the limits *before* we decide
whether it should be a deletion candidate.

This is an alternative to #1532 by @coolo - I think this way
of fixing it is clearer and better. It also fixes the tests.
@coolo couldn't figure out what was going on with the tests,
and I don't blame him, because it was quite tricky. In fact
what happened is that my understanding of the interaction
between the code and the assets used in the tests was slightly
wrong, but ironically, *this very bug* happened to exactly
'offset' my faulty understanding, so the tests I wrote happened
to pass. I have added some comments to hopefully explain this
correctly and in detail.

Signed-off-by: Adam Williamson <awilliam@redhat.com>