Add worker version check on the server.
If the worker have a lower version than the minimal version required by the
server the worker will disconnect and shutdown.

The worker's status message, now contains an extra parameter: The current
worker version.

This version is then updated in the database by the openqa server, and displayed
in the workers list.

Both versions (worker version, and the minimal required version) are
just a simple number and the version comparison number is just a simple
"different than" (!=).

Related: poo#7086, poo#26890, poo#10474