Remove assets from disk when deleting database entry (#1553)
See https://github.com/os-autoinst/openQA/pull/1551#issuecomment-353311600

Note: This changes the meaning of the DELETE route as it now actually deletes
the files on the file system. In don't think that this is a bad thing because
this way we can not forget to actually delete the file.

It also simplifies the GRU task code a little bit, because it
currently always removes the file from disk and then deletes the database
entry.