WebAPI: Add API description from Controller's POD
This commit adds methods into lib/OpenQA/WebAPI.pm to parse the Controller/API/V1
module files for POD documentation, and extract from there API descriptions to be
shown in the GUI's help page.

In order to parse POD, the module Tree::Pod was also added to the required modules
for the project.