Do not remove forwarded-for=https headers (#1558)
When I explained Ludwig how to avoid jobs expiring through mentioning in
bugzilla (and following the link), we noticed that it does not work for us.
Only after some experiments we realized that the 'report bug' template
puts http:// urls into bugzilla.opensuse.org. So most people's browsers
won't send refererrs (TLS default policy) and as such we have no linked
labels on opensuse.org

The reason is that we overwrite the forward-for=https header the login proxy
sends - as the apache on our openqa vm is not running on http, but gets
http requests from the login proxy.