Fedmsg plugin: emit 'standardized' fedmsg
There is a Red Hat-ish universe standard being established for
'CI'-ish message bus messages, the idea being that if many
different test systems emit messages in a common format, it
makes things much easier for other systems trying to follow
what all the test systems are doing. This makes the openQA
fedmsg plugin emit messages in this standardized format (at
least, in an initial approximation of it, with as-yet unapproved
extensions...) as well as the existing messages.

https://pagure.io/fedora-ci/messages