Kill all the children and grand children.
When a job finishes we need to be sure that we don't leave processes
running behind.

Since we don't want to loop the procfs to find grandchildren processes
we will use the systemd to kill all the groups in the cgroup.

The systemd will be trigger to kill all the cgroup processes when the
worker exits with an error status

Example of a process tree:

```
 worker(6288)───perl(6311)─┬─/usr/bin/isotov(6313)
                           ├─/usr/bin/isotov(6315)
                           └─/usr/bin/isotov(6317)─┬─qemu-system-x86(6321)─┬─{qemu-system-x86}(6324)
                                                   │                       ├─{qemu-system-x86}(6325)
                                                   │                       ├─{qemu-system-x86}(6326)
                                                   │                       └─{qemu-system-x86}(6328)
```

pgroups:
6288 - 6288
6311 - 6311
6313 - 6311
6315 - 6311
6317 - 6311
6321 - 6311
6318 - 6311
...