Emit openqa_job_create events when an ISO is posted (#1584)
We currently don't actually emit any openqa_job_create events
when posting an ISO - only a batch openqa_iso_create event. It
is easy to also emit an openqa_job_create event for each of the
newly-created jobs, so let's do that.

This also tweaks the Fedmsg plugin a bit, as this change exposed
an issue with it: when hundreds of events occur very quickly, as
now happens when an ISO is posted, the plugin seems to get backed
up waiting for the fedmsgs to be sent out. Eventually I think
something times out and the scheduler gets an error response,
which causes it to give up on scheduling the rest of the ISOs for
the compose it's working on. So to avoid that, we'll daemonize
the fedmsg-logger run so it doesn't block the plugin.

With this approach I suppose we actually run the risk of creating
too many fedmsg-logger processes simultaneously if we ever get to
the point where many hundreds or >1000 jobs are created for each
compose, but let's burn that bridge when we get to it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>