Allow saving needles for OS with dots in the version
The validation didn't allowed to save a needle for an OS
with a dot (.) in the name (for example: Opensuse leap 42.3)
if the VERSION was set.

The change in the validation allows for 1 dot, but not consecutive
dots.

The version is only used to find the location in the FS save the
needle.