Checks if failure status file exists before trying to slurp it
The status file that is created when an exception is launched isn't
immediatly available and the display of the details page doesn't
take that into consideration creating a race condition when someone
is following the test in the live tab, and then suddenly the test
have a hard failure.

The race condition is created, because the web interface will change
immediatly to the details page, and the exception information might
not be available.

Then a exception would be created and the interface is redirected to the
exception.html.ep template.

This solve this issue, because we don't need the information immediatly
available.