Some sync between code and tests
- catch errors during duplicating (tiny::catch got me again)
- restarting cluster jobs restarts the full cluster now. We had
  test cases that actually test the opposite of that.

parallel parents were not rescheduled when they were done. There
is no reasoning available in git history - it has always been this
way. But every test author I'm talking to either learned that you
have to trigger the parent or is confused that the clusters are
so touchy to which job you restarted