runcmd tests: handle differently-cased git error message
A git error message that 16-utils-runcmd.t expects to run into
has had its case changed in upstream git - formerly it was "Not
a git repository...", now it's "not a git repository...". This
was changed in fc045fe , "Mark messages for translations"; the
discussion on the upstream git mailing list indicates this is
to stay in line with a general policy on error messages for the
git project. As we're fairly likely to want this test to pass
with versions of git both before and after the change, let's
accept either form.

Signed-off-by: Adam Williamson <awilliam@redhat.com>