devel mode: Prevent endless, instant reconnects on error
Actually the JavaScript already waits 2 seconds before
reconnecting again when the error flag is set. However, the
error flag is cleared when a connection could be established.
Unfortunately that doesn't mean there's no more error. So
this changes the JavaScript to clear the error flag only
when we really received some status from os-autoinst.