Add global settings support for clone job
The global settings would be applied to all parental
jobs when a job is cloned if options skip-chained-deps
and --skip--deps are not specified.
Currently only 'WORKER_CLASS' is set as a global setting,
append others to be global if they can be set.

Also see: https://progress.opensuse.org/issues/25412