Use git -C for commits
Rather than specifying the git directory and work tree, use the -C
option so git switches to the specified directory and performs
operations. It's not required that the directory the files are being
added in contains the .git directory. Git will walk up the tree to find
the .git directory.

Requiring the .git directory in the commit directory effectively means
that commits can only be done in the root of the repository. Effectively
this means that the needles directory must be a git repo and not a
subdirectory of a git repo. Using -C allows that and lets git figure out
the repository structure.