fullstack/Docker test: Skip os-autoinst build only conditionally
Using CUSTOM_OS_AUTOINST is useful to prevent the `git clone` before
running the fullstack test inside docker. Skipping the build might
also be useful. However, it leads to problems if the provided build
is not compatible with the libraries found inside the container. So
it shouldn't be skipped by default.