Fix one more corner case for blocked_by
If a cluster depends on 2 independent parents, we need to make
sure to sort this out before looking at the cluster as a whole.
The situation is basically:

+------------+      +----------------+
|  Upgrade   |      |   Node 1       |
|  Node1     +------>                +---+------------------+
+------------+      +----------------+   | Supportserver    |
                                     |   |                  |
+------------+      +--------------------+------------------+
| Upgrade    |      |   Node 2       |
| Node 2     +------>                |
+------------+      +----------------+

And if you look at the cluster triggered by 'Upgrade Node1' you
might end up believing the whole cluster depends on 'Upgrade Node2',
while (of course) the Upgrades happen unblocked in parallel.

The cluster itself is blocked by both though