Fix removal of untracked assets
This is probably the most stupid bug ever - we basically only deleted untracked
assets by chance. delta_ms returns a Duration object - and that Duration object's
in_units only supports certain conversions. If the conversion is not supported,
0 is returned. So all these assets had basically all the time an age of 0

Using delta_days returns just that: the age in days and we happily remove them