Make sure multimachine jobs get a worker eventually
Increase a priority_offset whenever a job had to give up its
allocated worker due to other cluster jobs not getting one (starvation
in busy farms). Once this leads to a priority of 0 (or in case
the priority was 0 from the beginning), the worker won't be used
for other jobs so the cluster has a possibility to start

In case the cluster is too big for the worker farm, this will lead
to starvation in the other direction - but that's for the admin
to avoid. If we find this situation more often, we will need to
implement a timely degration of the priority_offset.