Handle `app` and `tx` attributes of Controller being weakened
In Mojolicious 8.03, the `tx` and `app` attributes of the
Controller class were weakened. This appears to mean that you
must set them using an object that's referenced elsewhere, or
else they just sort of...disappear, and the test blows up, with
rather cryptic errors about calling methods on undefined values.

Signed-off-by: Adam Williamson <awilliam@redhat.com>