Don't treat UEFI_PFLASH_VARS as asset if path is absolute
It seems we have two possible ways of using UEFI_PFLASH_VARS.
It may refer to the absolute filesystem location of a template
file, e.g. the ones shipped by edk2, or it may refer to an
image uploaded as an asset by a parent test (in the case that
we want a child test to 'inherit' the modifications to UEFI
vars that occurred during its parent's run).

We need to treat it as an asset when it really is an asset, so
it can be cached and found correctly - but we also need to *not*
treat it as an asset when it is not one, so openQA doesn't fail
to find it as an asset and cause the test to die. I think it
should be safe to simply go off whether the value looks like an
absolute path or not (i.e. starts with '/', since I don't think
anyone's doing this stuff on Windows...)

Signed-off-by: Adam Williamson <awilliam@redhat.com>