Don't leave a limit_assets task hanging after 14-grutasks.t
Since ea55d577, this subtest in grutasks creates a limit_assets
task, but doesn't run it. If the following subtest *does* run,
then it flushes the limit_assets task as well, and there's no
problem. But if the following subtest is skipped, because OBS_RUN
is set, then the limit_assets task just sorta hangs around in
gru's task queue...and if a later test runs gru, the limit_assets
task gets picked up and run. Because other tests do not mock out
the Assets schema `remove_from_disk` sub as grutasks does, this
means gru actually goes ahead and wipes out several asset files.
That may then cause another test to fail because it relies on
all the expected assets being present.

At present the above actually happens if the tests are run in
numerical order (e.g. with `prove t/`); when run with 'prove -r'
the bug is usually hidden as 14-grutasks happens to be run after
all other tests that run gru tasks, but we cannot rely on that
always being true.

Signed-off-by: Adam Williamson <awilliam@redhat.com>