Extend and improve asset documentation
Review of the previous commit ("Don't treat UEFI_PFLASH_VARS as
asset if path is absolute") exposed that we don't really have
great documentation on everything to do with asset handling -
it's quite short, vague and a little outdated. This is my
attempt to improve it, in my usual wordy style! It is more
precise about exactly what parameters are treated as assets,
explains in more detail what assets *are*, explains the deal
with UEFI_PFLASH_VARS, and explains 'storing' and 'publishing'
assets (since AFAICS this was not actually explained anywhere
in the docs at all).

Signed-off-by: Adam Williamson <awilliam@redhat.com>