21-needles.t: Use `discard_changes` instead of re-calling `find`
This test seems to fail quite often in Fedora package builds, on
the "last_updated changed after updating needle from editor"
check. I'm not 100% sure, but I *think* this may be because
the test's `$needle` is not reliably refreshed by this approach
of calling `find` on the resultset again. Changing it to use
discard_changes instead seems to help.

Signed-off-by: Adam Williamson <awilliam@redhat.com>