Revert support for ASSET_\d+_URL
For the time being seems that the dynamic schedule of tests with
caching seems a bit complicated. Mainly due to some tests using ASSET_\d
as part of an URL and others using it as a file to be expected to be
downloaded. The cache service should only treat assets as something to
download.

The feature introduced in #1855 is nice, but won't work with caching for
the time being. See https://progress.opensuse.org/issues/43511