Setting to not cancel parallel parents with still-pending children
As discussed extensively in
https://progress.opensuse.org/issues/46295 , openQA job logic
makes an assumption that, any time a parallel child fails or is
cancelled, its parent and any other pending children of that
parent ought to be cancelled. This is the behaviour SUSE's tests
expect, but it is not the behaviour Fedora's tests expect. In
Fedora we have several cases of clusters where a parallel parent
acts as a server to multiple unrelated child tests; if one of
the children fails, that does not mean the parent and all other
children must be cancelled.

This patch adds a job setting to set whether parallel parents
with other pending children (and hence those children) will be
cancelled when one child fails or is cancelled. The default is
the current behaviour. For the parent and the other pending
children *not* to be canceled, the parent must have the setting
`PARALLEL_CANCEL_WHOLE_CLUSTER` set to 0 (or anything false-y;
empty string also works).

Signed-off-by: Adam Williamson <awilliam@redhat.com>