worker: Fix "final status update"
The final status update is a bit special. We shouldn't apply the
usual error handling of stopping the job because that's from where
the status updated is called in that case. Instead we must ensure
to always actually invoke the callback. Otherwise stopping the job
can not proceed leaving the worker in an inconsistently cleaned
state.

See https://progress.opensuse.org/issues/48422