Don't delete gru task from db if minion job has not completed
In Fedora openQA, I noticed a problem related to the recent
change to run gru/minion tasks/jobs in parallel. We put guards
in download_asset and several other non-parallel-safe tasks to
prevent multiple instances of them running in parallel with
each other: the tasks use a 'guard' (basically a lock) and if
they cannot acquire it, they restart themselves.

However, it seems that when this happens, the minion execute()
returns and we wind up in this block of our superclassed
execute() which assumes that at this point the job has actually
run, and deletes the corresponding entry from the gru task db.
This was a big problem in Fedora, because we schedule the same
asset as multiple flavors, so each flavor gets its jobs blocked
on a download_asset gru task. One task executes 'normally' and
will not be deleted from the gru task db until the asset is
actually downloaded, but the other task hits the retry mechanism
and is immediately deleted from the db even though the asset has
not yet downloaded...so all the jobs for that flavor run, and
fail immediately.

To protect against this, we can check the state of the minion
job after we call `finish` and only delete the task from the db
if the job is actually 'failed' or 'finished'. Otherwise we
leave the entry in the db and wait for the job to run again and
actually complete this time.

Related ticket: [poo#48554](https://progress.opensuse.org/issues/48554)

Signed-off-by: Adam Williamson <awilliam@redhat.com>