AMQP plugin: optionally emit Fedora CI message spec messages too
This makes the AMQP publisher plugin emit messages in Fedora CI
message spec format, if a config key is set. See
https://pagure.io/fedora-ci/messages for more on this format.
The Fedmsg plugin already does this (with a slightly earlier
version of the format), but Fedora is migrating off ZMQ-based
fedmsg onto AMQP-based fedora-messaging, so I want to switch
Fedora from the Fedmsg plugin to this one.

Signed-off-by: Adam Williamson <awilliam@redhat.com>