docs: Combine "asset cleanup" with "asset handling"
"Asset Handling" should not be a minor point when "Asset Cleanup" is on a
higher level so this commit brings both sections together and therefore move
the "Developer Mode" section into the fitting "Use of the web UI" section.

"Asset Cleanup" is now a lower section within the "Asset Handling" section
which better introduces the overall concept and wording first before the
cleanup is described and how to configure it.