docs: Revamp the "job templates" section
The section "Using job templates to automate jobs creation" was part of
"GettingStarted" for a long time but did not really help beginners so this
commit moves it into the Users Guide.

The wording is updated to mention "job groups" as the place where the
test scenarios are actually defined. The necessary test definition steps are
updated to mention "job groups" in the correct place.

"Priorities" were previously incorrectly described with "test suites" and were
therefore moved to the section about "job groups" where they actually appear
in the web UI.

The order of sub-sections is adjusted according to the suggested order of
definition in the introductory section description.

The different ways how the test scenarios can be configured within the job
groups are described.