Revert to usual wording "distri" in YAML schedules
Despite we might think that spelling out words is better for
understanding we have to stick to our defined wording and also the API.
http://open.qa/docs/#concepts defines the word "distri" and we also use
that in os-autoinst as well as the actual test distributions, e.g.
os-autoinst-distri-opensuse. In case we really want to change that we
would need to handle all downstream uses as well.

With "prio" and "priority" we can more easily keep the change
because that value was in most cases only used internally
and only within openQA specifically and not os-autoinst.

Related progress issue: https://progress.opensuse.org/issues/46667