fetchneedles: Go back to old, quiet mode
For a shell script which does not support a "quiet" mode we should try
to stick to the *nix best practices of not outputtin anything unless it
is an error. The main reason for that is to not spam the cron account
running "fetchneedles" recurringly with unneeded cron output emails.