Provide style check feedback earlier with parallel execution
The test matrix part "FULLSTACK=0 UITESTS=0" was commonly the longest
running on travis CI already with about 23 minutes in comparison to the
next longest which is "FULLSTACK=0 UITESTS=1" with about 17 minutes.
Other parts take even less then that. This commit will conduct the style
checks in parallel to get earlier feedback on style issues as well as
save overall test runtime.