spec: Split out all dependencies properly per sub-package
The meaning of "t_requires" was probably not really clear to many so this commit
provides better names to explain the meaning. In the same go we split the
requirements per sub-package and pull in all runtime requirements as needed
into the build environment to execute the tests without pulling in more
runtime dependencies into the individual tickets than needed.