docs: Fix quotes and apostrophes for asciidoctor/github style
As observed on e.g.
http://open.qa/docs/#_show_previous_results_in_test_results_page_gh538
the quotes and apostrophes were rendered incorrectly by asciidoctor as
well as github in e.g.
https://github.com/os-autoinst/openQA/blob/master/docs/Installing.asciidoc#introduction
whereas locally using the tool "asciidoc" or pandoc – as used for
conversion in https://github.com/os-autoinst/openQA/pull/2127 – renders
fine with the `` style. This commit changes to the style preferred by
asciidoctor which we use to generate the documentation to fix the
rendering on our official documentation page.

https://github.com/asciidoc/asciidoc which is the upstream of the
openSUSE "asciidoc" package declares itself obsolete in favor of
asciidoctor so one more reason to go for that style :)