t/api/03-auth.t: don't actually wipe asset files from disk
In testing whether we're allowed to or forbidden from deleting
assets in various cases, we actually wind up deleting an asset.
This is a problem if t/ui/07-file.t runs after this test,
because one of its subtests relies on that asset being present.
This doesn't show up in Travis runs because the UI and API tests
are run in separate jobs, but it *does* show up in package
builds - you can see the test failing in OBS openQA package
build logs, for e.g.

Signed-off-by: Adam Williamson <awilliam@redhat.com>