AMQP: allow passing headers to publish_amqp
Fedora's new AMQP-based messaging system requires messages to
have certain headers:

https://fedora-messaging.readthedocs.io/en/stable/wire-format.html

I'm planning to write a plugin for emitting Fedora-compliant
AMQP messages as a subclass of the existing AMQP plugin, but in
order to be able to use `publish_amqp` in that plugin, I need to
be able to pass it headers like this. Otherwise I'd have to
reimplement it.

This implementation aims to ensure it works just as before if
you *don't* want to pass it headers, and without needing to
duplicate the `publish_p` calls for "headers" and "no headers"
cases - that's the reason for making `$headers` a reference to
an empty hash if it's not defined, and then checking that it's
a hashref: to be sure about what we're asking `publish_p` to do.
Passing it a ref to an empty hash should effectively be a no-op
so far as headers are concerned, and not interfere with passing
`routing_key` as a param.

Signed-off-by: Adam Williamson <awilliam@redhat.com>