load_templates: fix --clean and reading perl templates
This fixes two bugs in load_templates since #2204 landed. First,
`--clean` couldn't work once JobGroups was added to `@tables`
because that API endpoint returns an array (not a hash) and the
script would choke on that (it only knows how to handle hashes).
The correct fix is just to skip JobGroups when doing `--clean`
because we usually cannot delete them anyway, as non-empty job
groups cannot be deleted. There's no point trying to handle the
array.

Second, loading perl format template files was broken by the
change to make `$datafile` the slurped contents of the file
rather than simply the filename. We have to change `do` to `eval`
to handle that change.

Signed-off-by: Adam Williamson <awilliam@redhat.com>