apparmor: Allow loading of client.conf in usual home of geekotest
This fixes the "Permission denied" problem as observed on o3 when openQA
tries to access the client.conf file in the home directory of user
"geekotest". While this file is not directly used by openQA but instead
/etc/openqa/client.conf the file in the home directory is necessary for
the sync and trigger scripts currently.

Related progress issue: https://progress.opensuse.org/issues/54308