Validate tag input field in needleeditor
Currently, it is possible to add a tag with spaces or other arbitrary
characters. Spaces are only inconvenient, as they are only visible when
inspecting the generated JSON, other characters may have worse effects.

Use the HTML5 "pattern" attribute to check the value, and reflect the
state in the form field background (light red) and submit button.