Rely on t_updated for the worker's online status in the web UI
This should effectively yield the same result because
t_updated is updated accordingly when the worker

* connects to the ws server.
* disconnects from the ws server (t_updated is decreased).
* reports its status via ws connection.
* updates the job status via the REST API.

Note that previously the worker's offline status was deduced
by checking the status of the Mojo::Transaction representing
the websocket connection to the worker. That meant that previously
the inactivity timeout for web socket connections decided after
which time of unresponsiveness we consider a worker offline.
Now that is only defined by the WORKERS_CHECKER_THRESHOLD constant.