Recompute worker's broken state after job finished
The periodic recompution of the worker's brokenness is not done
during job execution (because then the status is always 'working'
and never 'broken'). That means without this change the broken
status might not be recomputed before accepting the next job and
a worker might keep busy producing incomplete jobs.

See https://progress.opensuse.org/issues/56993