Prevent potential deadlocks in scheduler and ws server
This prevents the scheduler and ws server from querying itself
which would lead to a deadlock (see
https://progress.opensuse.org/issues/57017).

This change makes problems like this easier to debug by letting
the scheduler and ws server die in that situation with a clear
error message.

OpenQA::Schema::Result::Workers::send_command is still handling
the case that it is being called from the ws server itself
because that function is called all over the place and I like to
keep it robust. However, the hack to use the application's name
for that check has been removed.